add component profile popup
